tiktok-automation/requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
python-dotenv==1.0.0